node-main/deps/v8/test/webkit/fast/js/Promise-static-reject-expected.txt

Test Promise.reject

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is undefined
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
